pt: fix type mismatch

+++ ftnoir_tracker_pt/point_extractor.cpp
@@ -131,7 +131,7 @@ std::vector<cv::Vec2f> PointExtractor::extract_points(cv::Mat& frame)
                     ret.push_back(b_);
 
                 char buf[64];
-                sprintf(buf, "%d%% %d px", b_.pos.size()*100/successes, (int)(2.*sqrt(b_.effective_area()) / sqrt(3.14)));
+                sprintf(buf, "%d%% %d px", (int)(b_.pos.size()*100/successes), (int)(2.*sqrt(b_.effective_area()) / sqrt(3.14)));
                 const auto pos = b_.effective_pos();
                 cv::putText(frame, buf, cv::Point(pos[0]+30, pos[1]+20), cv::FONT_HERSHEY_DUPLEX, 1, cv::Scalar(0, 0, 255), 1);
             }
